added support for better error handling

+++ gestion-laboratoire/src/main/java/com/example/gestion_laboratoire/services/LaboratoireService.java
@@ -1,14 +1,17 @@
 package com.example.gestion_laboratoire.services;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.example.gestion_laboratoire.Models.Laboratoire;
+import com.example.gestion_laboratoire.models.Laboratoire;
 import com.example.gestion_laboratoire.repositories.LaboratoireRepository;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class LaboratoireService {
@@ -26,33 +29,33 @@ public class LaboratoireService {
         return laboratoireRepository.findAll();
     }
 
-    public Laboratoire getLaboratoiresByIdLong(Long id) throws Exception {
+    public Laboratoire getLaboratoiresByIdLong(Long id) throws EntityNotFoundException {
         if (laboratoireRepository.existsById(id))
             return laboratoireRepository.findById(id).get();
         else
-            throw new Exception("Laboratoire Not found");
+            throw new EntityNotFoundException("Laboratory Not found");
     }
 
-    public String createLaboratoire(Laboratoire laboratoire) {
+    public ResponseEntity<Object> createLaboratoire(Laboratoire laboratoire) {
         try {
-            System.out.println(laboratoire);
             String imageName = laboratoire.getNom() + "_" + laboratoire.getId();
             String imageURL = cloudinaryService.uploadImage(imageName, laboratoire.getImageFile());
             if (imageURL == "Failed to upload image") {
-                return imageURL;
+                return new ResponseEntity<>("Could not create laboratory : " + imageURL, HttpStatus.FAILED_DEPENDENCY);
             }
             laboratoire.setLogo(imageURL);
             laboratoireRepository.save(laboratoire);
-            return "Laboratoire created successfully";
-        } catch (IllegalArgumentException ie) {
-            return "Unknown error has occured during the creation of the Laboratoire";
+            return new ResponseEntity<>("Laboratory created successfully", HttpStatus.CREATED);
+        } catch (Exception ie) {
+            return new ResponseEntity<>("Unknown error has occured during the creation of the Laboratory",
+                    HttpStatus.BAD_REQUEST);
         }
     }
 
     @Transactional
-    public String updateLaboratoire(Long id, Laboratoire laboratoire) {
+    public ResponseEntity<Object> updateLaboratoire(Long id, Laboratoire laboratoire) {
         System.out.println(laboratoire);
-        try {
+        if (laboratoireRepository.existsById(id)) {
             Laboratoire labo = laboratoireRepository.findById(id).get();
             if (laboratoire.getNom() != null && laboratoire.getNom().length() != 0)
                 labo.setNom(laboratoire.getNom());
@@ -65,21 +68,25 @@ public class LaboratoireService {
             if (laboratoire.getImageFile() != null && laboratoire.getImageFile().length != 0)
                 laboratoire.setLogo(cloudinaryService.uploadImage(laboratoire.getNom() + "_" + laboratoire.getId(),
                         laboratoire.getImageFile()));
-            return "Laboratoire " + laboratoire.getNom() + " updated";
-
-        } catch (Exception e) {
-            return "laboratoire Not Found";
+            return new ResponseEntity<>("Laboratory " + laboratoire.getNom() + " updated", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Laboratory not found", HttpStatus.NOT_FOUND);
         }
     }
 
-    public String deleteLaboratoire(Long id) {
-        try {
+    public ResponseEntity<Object> deleteLaboratoire(Long id) {
+        if (laboratoireRepository.existsById(id)) {
             Laboratoire laboratoire = laboratoireRepository.findById(id).get();
-            cloudinaryService.deleteImage(laboratoire.getNom() + "_" + laboratoire.getId());
+            String imageDeletionMessage = cloudinaryService
+                    .deleteImage(laboratoire.getNom() + "_" + laboratoire.getId());
+            if (imageDeletionMessage != "Image deleted successfully") {
+                return new ResponseEntity<>("Could not delete laboratory, " + imageDeletionMessage,
+                        HttpStatus.FAILED_DEPENDENCY);
+            }
             laboratoireRepository.deleteById(id);
-            return "Laboratoire " + laboratoire.getNom() + " deleted";
-        } catch (NoSuchElementException e) {
-            return "laboratoire Not Found";
+            return new ResponseEntity<>("Laboratory " + laboratoire.getNom() + " deleted", HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>("Laboratory not found", HttpStatus.NOT_FOUND);
         }
     }
